requirements.txt
streamlit>=1.32.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
mysql-connector-python>=8.0.0
scipy>=1.10.0
